refactor(auth): rename Auth to NixAuth and move to nix_auth module
Rename providers::auth to providers::nix_auth and the Auth struct to
NixAuth, to clarify its purpose alongside the new git_auth module.
This module handles netrc-based authentication for nix operations
(nix copy, store access), while git_auth handles git transport auth.

Pure rename and import updates — no behavioral changes.
